Having ignore files to ignore some sources from HAPI FHIR directories and config files.
Overlay project depends on some web sources. We will not be working on
those sources. Thus, they need to be manually placed in the local
workspace. This ignore file will ignore to be added to the repository.

database config file contains username and password. This will be
ignored as well.